requirements.txt
Django==5.0.2
djangorestframework==3.14.0
uWSGI==2.0.22
django-cors-headers==4.3.1
mysqlclient==2.2.1
djangorestframework-simplejwt==5.3.1